ai-router/requirements.txt
requests
PyYAML
evdev
